ipl/requirements.txt
Django>=2.0,<2.1
psycopg2-binary
django-colorfield
simplejson
XlsxWriter
uwsgi
python-dateutil
Pillow
django-grappelli<=2.11.2
git+https://github.com/herzog31/django-filebrowser.git#egg=django-filebrowser
tablib